Refactor: View 주석 달기, 파일 class 분리
