+++ BackEnd/model/__init__.py
@@ -1,7 +1,9 @@
-from model.order_dao   import CartDao, OrderDao, ShipmentDao
-from model.product_dao import ProductDao
-from model.user_dao    import AccountDao
-from model.util_dao    import SelectNowDao
+from model.order_dao    import OrderDao
+from model.cart_dao     import CartDao
+from model.product_dao  import ProductDao
+from model.user_dao     import AccountDao
+from model.shipment_dao import ShipmentDao
+from model.util_dao     import SelectNowDao
 
 __all__ = [
     "CartDao",
