+++ BackEnd/model/product_dao.py
@@ -65,7 +65,7 @@ class ProductDao:
             product_list = cursor.fetchall()
             return product_list
 
-    def get_product_detail(self, filters, connection):
+    def get_product_detail(self, connection, filters):
 
         query = """
             SELECT
@@ -110,7 +110,7 @@ class ProductDao:
 
             return result
 
-    def get_product_image(self, filters, connection):
+    def get_product_image(self, connection, filters):
 
         query = """
             SELECT     
@@ -130,7 +130,7 @@ class ProductDao:
 
             return result
     
-    def get_product_option_information(self, filters, connection):
+    def get_product_option_information(self, connection, filters):
         
         query = """
             SELECT
@@ -153,5 +153,171 @@ class ProductDao:
             cursor.execute(query, filters)
 
             result = cursor.fetchall()
+
+            return result
+    
+    def product_exist_check(self, connection, filters):
+        """상품 존재 여부 확인
+
+        Author:
+            백승찬
+
+        Args:
+            filters (dict): cart_id 또는 product_option_id 값을 가지는 dictionary
+            connection (객체): pymysql 객체
+
+        Raises:
+
+        Returns:
+            200 : product_history_id
+        """
+
+        query = """
+            SELECT
+                ph.id
+            
+            FROM
+                product_histories AS ph
+        """
+
+        if filters.get("cart_id"):
+            query += f"""
+                INNER JOIN carts AS c
+                        ON c.product_id = ph.product_id
+
+                WHERE
+                    c.id = %(cart_id)s
+                    AND ph.end_time = "{END_DATE}"
+                    AND ph.is_deleted = false
+                    AND ph.is_sold = true
+            """
+
+        if filters.get("product_option_id"):
+            query += f"""
+                INNER JOIN product_options AS po
+                        ON po.product_id = ph.product_id
+
+                WHERE
+                    po.id             = %(product_option_id)s
+                    AND ph.end_time   = "{END_DATE}"
+                    AND ph.is_deleted = false
+                    AND ph.is_sold    = true
+            """
+        
+        with connection.cursor(pymysql.cursors.DictCursor) as cursor:
+
+            cursor.execute(query, filters)
+
+            result = cursor.fetchone()
+
+            return result
+        
+    def product_option_sold_out_check(self, connection, filters):
+        """상품 옵션 sold_out 확인
+
+        Author:
+            백승찬
+
+        Args:
+            filters (dict): cart_id 또는 product_option_id 값을 가지는 dictionary
+            connection (객체): pymysql 객체
+
+        Raises:
+
+        Returns:
+            200 : product_option_id
+        """
+
+        query = """
+            SELECT 
+                po.id
+            FROM
+                product_options AS po
+        """
+
+        if filters.get("cart_id"):
+            query += """
+                INNER JOIN carts AS c
+                        ON c.product_option_id = po.id
+                WHERE
+                    c.id = %(cart_id)s
+                    AND po.is_sold_out = false
+            """
+        
+        if filters.get("product_option_id"):
+            query += """
+                WHERE
+                    po.id = %(product_option_id)s
+                    AND po.is_sold_out = false
+            """
+        
+        with connection.cursor(pymysql.cursors.DictCursor) as cursor:
+
+            cursor.execute(query, filters)
+
+            result = cursor.fetchone()
+
+            return result
+    
+    def get_discount_product_check(self, connection, filters):
+
+        query = f"""
+            SELECT
+                ph.price - TRUNCATE((ph.price * (ph.discount_rate/100)), -1) AS sale_price
+
+            FROM product_histories AS ph
+
+            WHERE
+                ph.product_id     = %(product_id)s
+                AND ph.is_sold    = true
+                AND ph.end_time   = "{END_DATE}"
+                AND ph.is_deleted = false
+        """
+
+        with connection.cursor(pymysql.cursors.DictCursor) as cursor:
+
+            cursor.execute(query, filters)
+
+            result = cursor.fetchone()
+
+            return result
+    
+    def update_product_option_stock(self, connection, filters):
+
+        query = """
+            UPDATE
+                product_options
+
+            SET
+                stock = stock - %(quantity)s
+
+            WHERE
+                id = %(product_option_id)s
+                AND stock - %(quantity)s > -11
+        """
+
+        with connection.cursor() as cursor:
+
+            result = cursor.execute(query, filters)
+
+            return result
+    
+    def update_product_option_is_sold_out(self, connection, filters):
+
+        query = """
+            UPDATE
+                product_options
+
+            SET
+                is_sold_out = true
+
+            WHERE
+                id        = %(product_option_id)s
+                AND stock = -10
+        """
+
+        with connection.cursor() as cursor:
+
+            result = cursor.execute(query, filters)
 
             return result
