+++ BackEnd/service/__init__.py
@@ -1,12 +1,13 @@
-from service.order_service   import CartService, OrderService, ShipmentService, OrderCompleteService
-from service.product_service import ProductService
-from service.user_service    import SignInService
+from service.order_service    import OrderService
+from service.cart_service     import CartService
+from service.shipment_service import ShipmentService
+from service.product_service  import ProductService
+from service.user_service     import SignInService
 
 __all__ = [
     "CartService",
     "OrderService",
     "ShipmentService",
     "ProductService",
-    "SignInService",
-    "OrderCompleteService"
+    "SignInService"
 ]
