+++ BackEnd/view/product_view.py
@@ -5,7 +5,6 @@ from flask_request_validator import GET, PATH, Param, validate_params, Composite
 from service    import ProductService
 from connection import connect_db
 
-
 class ProductView(MethodView):
 
     # 로그인 여부 데코레이터
@@ -46,39 +45,6 @@ class ProductView(MethodView):
         except Exception as e:
             raise e
 
-        finally:
-            if connection is not None:
-                connection.close()
-    
-class ProductDetailView(MethodView):
-
-    @validate_params (
-        Param("product_id", PATH, int),
-        Param("offset", GET, int, required=False),
-        Param("limit", GET, int, required=False)
-    )
-    def get(*args, product_id):
-
-        filters = {
-            "offset"     : int(request.args.get("offset", 0)),
-            "limit"      : int(request.args.get("limit", 5)),
-            "product_id" : int(product_id)
-        }
-
-        product_service = ProductService()
-        connection      = None
-
-        try:
-            connection = connect_db()
-            result     = product_service.get_product_detail_list(filters, connection)
-
-            return jsonify({"data": result})
-
-        except Exception as e:
-            if connection is not None:
-                connection.rollback()       
-                raise e
-        
         finally:
             if connection is not None:
                 connection.close()
